fix typescript errors
